requirements-test.txt
# Testing dependencies for StudentCBT

# Core testing framework
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# For Excel file testing
openpyxl==3.1.2

# For better test output
pytest-html==4.1.1
pytest-sugar==0.9.7

# For parallel test execution
pytest-xdist==3.5.0

# For mocking and fixtures
pytest-mock==3.12.0
faker==21.0.0

# Code coverage
coverage==7.4.0
